refactor: utils.js 파일 auth 폴더로 이동
