+++ insta-clone/src/pages/Auth/Signup.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 import logoSvg from '/assets/icons/logo.svg';
-import validateAuth from '@/utils/utils';
+import validateAuth from '@/pages/Auth/utils';
 import apiClient from '@/services/apiClient';
 
 const Signup = () => {
